requirements.txt
pdfplumber==0.10.3
python-docx==1.0.1
streamlit==1.32.0 